:recycle: Refactor: 서버 반환값에 맞게 데이터 타입 수정

+++ app/src/main/java/com/example/snack4diet/api/getBookmark/Data.kt
@@ -3,6 +3,6 @@ package com.example.snack4diet.api.getBookmark
 data class Data(
     val baseNutrition: BaseNutrition,
     val count: Int,
-    val favoriteFoodId: Int,
+    val favoriteFoodId: Long,
     val name: String
 )
